Load shaders from their own files
This enables easier editing and reading of the shaders, compared to
inline strings. It also enables iterative development of the shaders
without requiring a rebuild of the program.

The CMake custom target and rule have a set of desired and undesired
behaviors. We've chosen to implement the file copy to meet these
requirements.

Desired behavior:
 - Copies whenever the shaders are modified, during the build stage.

Undesired behavior:
 - Only copies when hello_gui target is built
 - Only copies during config stage
 - Doesn't copy when shader is modified unless hello_gui also rebuilds

These behaviors preclude using file() or configure_file() (which only
copies during config stage) or adding the custom command to hello_gui
(which only updates if a dependency of hello_gui updates, and we don't
want hello_gui to have a dependency on the shaders because the C/C++
code doesn't need rebuilt if the shader changes).